Change: type variables can now only be single letters, same as free vars. Error on unknown types in restrictions (on instantiation). Fixes #487, and it actually uncovered a few bugs/mistakes.

+++ spec/compiler/type_inference/def_spec.cr
@@ -119,7 +119,7 @@ describe "Type inference: def" do
       def foo(x : Int, y : Int)
       end
 
-      def foo(x : Int, y : Double)
+      def foo(x : Int, y : Float)
       end
 
       foo(1 || 'a', 1 || 1.5)
